+++ src/main/java/com/andersmmg/create_alerted/AllSoundEvents.java
@@ -10,15 +10,6 @@ import java.util.function.Supplier;
 public class AllSoundEvents {
     public static final DeferredRegister<SoundEvent> SOUND_EVENTS = DeferredRegister.create(Registries.SOUND_EVENT, CreateAlerted.MODID);
 
-    public static final Supplier<SoundEvent> ALARM_BASIC = SOUND_EVENTS.register("alarm_basic",
-            () -> SoundEvent.createVariableRangeEvent(ResourceLocation.fromNamespaceAndPath(CreateAlerted.MODID, "alarm_basic")));
-
-    public static final Supplier<SoundEvent> ALARM_ANNOYING = SOUND_EVENTS.register("alarm_annoying",
-            () -> SoundEvent.createVariableRangeEvent(ResourceLocation.fromNamespaceAndPath(CreateAlerted.MODID, "alarm_annoying")));
-
-    public static final Supplier<SoundEvent> ALARM_BUZZ = SOUND_EVENTS.register("alarm_buzz",
-            () -> SoundEvent.createVariableRangeEvent(ResourceLocation.fromNamespaceAndPath(CreateAlerted.MODID, "alarm_buzz")));
-
     public static final Supplier<SoundEvent> SMOKE_DETECTOR = SOUND_EVENTS.register("smoke_detector",
             () -> SoundEvent.createVariableRangeEvent(ResourceLocation.fromNamespaceAndPath(CreateAlerted.MODID, "smoke_detector")));
 }