+++ src/main/java/com/andersmmg/create_alerted/block/AlarmType.java
@@ -3,12 +3,15 @@ package com.andersmmg.create_alerted.block;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.resources.ResourceLocation;
+import org.jetbrains.annotations.Nullable;
 
-public record AlarmType(ResourceLocation soundId, int interval) {
+import java.util.Optional;
+
+public record AlarmType(@Nullable ResourceLocation soundId, @Nullable Integer interval) {
     public static final Codec<AlarmType> CODEC = RecordCodecBuilder.create(instance ->
             instance.group(
-                    ResourceLocation.CODEC.fieldOf("sound").forGetter(AlarmType::soundId),
-                    Codec.INT.fieldOf("interval").forGetter(AlarmType::interval)
-            ).apply(instance, AlarmType::new)
+                    ResourceLocation.CODEC.optionalFieldOf("sound").forGetter(type -> Optional.ofNullable(type.soundId())),
+                    Codec.INT.optionalFieldOf("interval").forGetter(type -> Optional.ofNullable(type.interval()))
+            ).apply(instance, (sound, interval) -> new AlarmType(sound.orElse(null), interval.orElse(null)))
     );
 }